Extract table of differential results from 'K2' object

Create table differential analysis results from 'K2' object.

getDGETable(K2res)

Arguments

K2res
An object of class K2.

Value

A data.frame object with the following columns:

gene: The gene ids from the `ExpressionSet` object

coef: The difference between the means of each subgroup at a given partition

mean: The mean across all observations at the given partition

t: The test statistic estimated by the `limma` R package

pval: The p-value estimated by the `limma` R package

fdr: The multiple hypothesis corrected fdr p-value, adjusted across all partitions

B: The B-statistic estimated by the `limma` R package

edge: Indication of which subgroup the gene was assigned at a given partition

node: The identifier of the partition

direction: The direction of coefficient for the assigned gene

References

Reed ER, Monti S (2020). “Multi-resolution characterization of molecular taxonomies in bulk and single-cell transcriptomics data.” Bioinformatics. doi: 10.1101/2020.11.05.370197 , http://biorxiv.org/lookup/doi/10.1101/2020.11.05.370197. Ritchie ME, Phipson B, Wu D, Hu Y, Law CW, Shi W, Smyth GK (2015). “limma powers differential expression analyses for RNA-sequencing and microarray studies.” Nucleic Acids Research, 43(7), e47--e47. ISSN 1362-4962, 0305-1048, doi: 10.1093/nar/gkv007 , http://academic.oup.com/nar/article/43/7/e47/2414268/limma-powers-differential-expression-analyses-for. Benjamini Y, Hochberg Y (1995). “Controlling the False Discovery Rate: A Practical and Powerful Approach to Multiple Testing.” Journal of the Royal Statistical Society: Series B (Methodological), 57(1), 289--300. ISSN 00359246, doi: 10.1111/j.2517-6161.1995.tb02031.x , http://doi.wiley.com/10.1111/j.2517-6161.1995.tb02031.x.

Examples

## Read in ExpressionSet object
library(Biobase)
data(sample.ExpressionSet)

## Pre-process and create K2 object
K2res <- K2preproc(sample.ExpressionSet)

## Run K2 Taxonomer algorithm
K2res <- K2tax(K2res,
               stabThresh = 0.5)

## Run differential analysis on each partition
K2res <- runDGEmods(K2res)

## Extract table of differential results
head(getDGETable(K2res))

#>       gene     coef      mean          t         pval          fdr         B
#> 1 31737_at 1445.129  264.7562 119.688475 9.341821e-10 4.203820e-06 -4.445513
#> 2 31583_at 1519.998 2819.3481   8.373340 1.046179e-08 2.353902e-05 -2.345377
#> 3 31546_at 1242.405 1963.0257   8.138876 1.768093e-08 2.652140e-05 -2.384707
#> 4 31573_at 1229.790 2034.8062   7.396055 9.781121e-08 1.100376e-04 -2.523671
#> 5 31511_at 1300.542 3256.3362   7.292037 1.250010e-07 1.125009e-04 -2.545017
#> 6 31568_at 1422.450 2866.9177   7.147891 1.760053e-07 1.320040e-04 -2.575418
#>   edge node direction
#> 1    2    I        up
#> 2    2    A        up
#> 3    2    A        up
#> 4    2    A        up
#> 5    2    A        up
#> 6    2    A        up