Create interactive table of gene set enrichment results from 'K2' object

Create table gene set enrichment results from 'K2' object.

getEnrichmentInter(
  K2res,
  maxFDR_score = 0.01,
  minDiff_score = NULL,
  maxPval_score = NULL,
  maxFDR_fisher = NULL,
  maxPval_fisher = NULL,
  gsNames = NULL,
  nodes = NULL,
  pagelength = 50
)

Arguments

K2res

A K2 class object.

maxFDR_score

Numeric. A value between 0 and 1 indicating the FDR cutoff for differential analysis of enrichment scores.

minDiff_score

Numeric. A value between 0 and 1 indicating the mean difference cutoff for differential analysis of enrichment scores.

maxPval_score

Numeric. A value between 0 and 1 indicating the p-value cutoff for differential analysis of enrichment scores.

maxFDR_fisher

Numeric. A value between 0 and 1 indicating the FDR cutoff for Fisher-based overrepresentation analysis.

maxPval_fisher

Numeric. A value between 0 and 1 indicating the p-value cutoff for Fisher-based overrepresentation analysis.

gsNames

Character. A vector of gene sets identifiers to display.

nodes

Character. A vector of node identifiers to display.

pagelength

Numeric. Number of rows to display in each page of output.

Value

An interactive data frame with the following columns:

Gene Set: The identifier of the gene set.

Node: The identifier of the partition

Edge: The identifier of partition subgroup

Direction: The direction of coefficient for the assigned gene sets

P Value Fisher: The p-value estimated by Fisher-based overrepresentation analysis

FDR Fisher: The multiple hypothesis corrected FDR Fisher p-value, adjusted across all partitions

N Overlap: The intersection of differentially expressed genes and gene set

N Sig. Genes: The number differentially expressed genes

N Gene Set: The number genes comprising the gene set

P Value Score: The p-value estimated by differential analysis of enrichment scores

FDR Score: The multiple hypothesis corrected FDR p-value estimated by differential analysis of enrichment scores, adjusted across all partitions

Diff Score: The difference between the mean enrichment score of each subgroup at a given partition