Perform tests to test for associations between subgroups and phenotypic variables

Adds statistical tests results to the output of K2tax() based on numeric and factor variables in info.

runTestsMods(K2res, infoClass = NULL)

Arguments

K2res
An object of class K2. The output of K2tax().

infoClass
A named vector denoted types of tests to run on phenotypic variables (See K2preproc()).

Value

An object of class K2.

Examples

## Read in ExpressionSet object
library(Biobase)
data(sample.ExpressionSet)

## Pre-process and create K2 object
K2res <- K2preproc(sample.ExpressionSet)

## Run K2 Taxonomer algorithm
K2res <- K2tax(K2res,
            stabThresh=0.5)

infoClassVector <- c(sex='factor', score='numeric1')

K2res <- runTestsMods(K2res, infoClass=infoClassVector)

head(getTestsModTable(K2res))

#>    value node edge      pval       fdr        stat df   obsMean   altMean
#> 9    sex    C    1 0.1088690 0.8307692   0.2223068 NA        NA        NA
#> 29 score    H    1 0.1277138 0.8307692  60.5000000 NA 0.6700000 0.5127273
#> 17 score    E    1 0.1458790 0.8307692 100.5000000 NA 0.6000000 0.4975000
#> 5    sex    B    1 0.1973244 0.8307692   4.7256460 NA        NA        NA
#> 18   sex    E    1 0.2279715 0.8307692   0.3183362 NA        NA        NA
#> 10 score    C    1 0.2417236 0.8307692  96.5000000 NA 0.5745455 0.5093333
#>      diffMean nhits ncase nalt ndrawn hits                         test
#> 9          NA    NA    NA   NA     NA   NA two-sided Fishers Exact Test
#> 29 0.15727273    NA    NA   NA     NA   NA          1-sided Wilcox Test
#> 17 0.10250000    NA    NA   NA     NA   NA          1-sided Wilcox Test
#> 5          NA    NA    NA   NA     NA   NA two-sided Fishers Exact Test
#> 18         NA    NA    NA   NA     NA   NA two-sided Fishers Exact Test
#> 10 0.06521212    NA    NA   NA     NA   NA          1-sided Wilcox Test