Named list of genesets used in K2Taxonomer run

Retrieve or assign named list of genesets for K2Taxonomer run.

K2genesets(K2res)

K2genesets(K2res) <- value

# S4 method for K2
K2genesets(K2res)

# S4 method for K2
K2genesets(K2res) <- value

Arguments

K2res
K2 class object.

value
Named list. A list with an element for each ge

Value

Named list. A list with an element for each gene set.

Examples

data(K2res)
head(K2genesets(K2res))

#> $GS1
#>  [1] "31737_at"                 "31583_at"                
#>  [3] "31546_at"                 "31573_at"                
#>  [5] "31568_at"                 "31511_at"                
#>  [7] "AFFX-HUMTFRR/M11507_3_at" "31397_at"                
#>  [9] "31589_at"                 "31596_f_at"              
#> [11] "31607_at"                 "31527_at"                
#> [13] "31545_at"                 "AFFX-PheX-3_at"          
#> [15] "31492_at"                 "31708_at"                
#> [17] "31330_at"                 "31564_at"                
#> [19] "AFFX-BioC-3_st"           "31509_at"                
#> [21] "AFFX-HSAC07/X00351_5_at"  "31719_at"                
#> [23] "31686_at"                 "31385_at"                
#> [25] "AFFX-HSAC07/X00351_M_at"  "31627_f_at"              
#> [27] "AFFX-MurIL2_at"           "31315_at"                
#> [29] "31726_at"                 "31502_at"                
#> [31] "31396_r_at"               "31556_at"                
#> [33] "31425_g_at"               "31574_i_at"              
#> [35] "31507_at"                 "31649_at"                
#> [37] "31550_at"                 "31590_g_at"              
#> [39] "31697_s_at"               "31571_at"                
#> [41] "31538_at"                 "31488_s_at"              
#> [43] "31580_at"                 "AFFX-PheX-M_at"          
#> [45] "31586_f_at"               "31477_at"                
#> [47] "31722_at"                 "31535_i_at"              
#> [49] "AFFX-YEL021w/URA3_at"     "31721_at"                
#> 
#> $GS2
#>  [1] "31461_at"                   "31579_at"                  
#>  [3] "31670_s_at"                 "31575_f_at"                
#>  [5] "31404_at"                   "AFFX-TrpnX-5_at"           
#>  [7] "31554_at"                   "AFFX-HUMGAPDH/M33197_3_at" 
#>  [9] "31543_at"                   "AFFX-YEL024w/RIP1_at"      
#> [11] "AFFX-HUMTFRR/M11507_M_at"   "31622_f_at"                
#> [13] "31505_at"                   "31463_s_at"                
#> [15] "31694_at"                   "31529_at"                  
#> [17] "31431_at"                   "31406_at"                  
#> [19] "31349_at"                   "AFFX-HUMISGF3A/M97935_3_at"
#> [21] "31537_at"                   "31684_at"                  
#> [23] "31728_at"                   "31386_at"                  
#> [25] "31569_at"                   "31618_at"                  
#> [27] "31560_at"                   "31417_at"                  
#> [29] "AFFX-BioB-5_at"             "31471_at"                  
#> [31] "31675_s_at"                 "31481_s_at"                
#> [33] "31623_f_at"                 "31420_at"                  
#> [35] "31466_at"                   "31364_i_at"                
#> [37] "31345_at"                   "31441_at"                  
#> [39] "31460_f_at"                 "AFFX-HUMGAPDH/M33197_5_at" 
#> [41] "AFFX-MurIL4_at"             "AFFX-HUMGAPDH/M33197_M_at" 
#> [43] "31631_f_at"                 "31371_at"                  
#> [45] "31512_at"                   "AFFX-CreX-5_at"            
#> [47] "31605_at"                   "31557_at"                  
#> [49] "AFFX-HUMGAPDH/M33197_3_st"  "31428_at"                  
#> 
#> $GS3
#>  [1] "31584_at"                    "31379_at"                   
#>  [3] "31635_g_at"                  "31320_at"                   
#>  [5] "31423_at"                    "31418_at"                   
#>  [7] "31335_at"                    "31578_at"                   
#>  [9] "31452_at"                    "31465_g_at"                 
#> [11] "31691_g_at"                  "31408_at"                   
#> [13] "31604_at"                    "31692_at"                   
#> [15] "31324_at"                    "31566_at"                   
#> [17] "31499_s_at"                  "31713_s_at"                 
#> [19] "AFFX-BioB-3_st"              "31720_s_at"                 
#> [21] "31671_at"                    "31593_at"                   
#> [23] "31711_at"                    "31530_at"                   
#> [25] "31520_at"                    "31416_at"                   
#> [27] "31653_at"                    "AFFX-BioC-3_at"             
#> [29] "31321_at"                    "31665_s_at"                 
#> [31] "AFFX-HSAC07/X00351_3_st"     "31508_at"                   
#> [33] "31380_at"                    "31392_r_at"                 
#> [35] "31506_s_at"                  "31672_g_at"                 
#> [37] "31534_at"                    "31419_r_at"                 
#> [39] "31387_at"                    "AFFX-HUMISGF3A/M97935_MA_at"
#> [41] "31432_g_at"                  "31427_at"                   
#> [43] "31328_at"                    "31389_at"                   
#> [45] "AFFX-TrpnX-M_at"             "31519_f_at"                 
#> [47] "31685_at"                    "31648_at"                   
#> [49] "31438_s_at"                  "31680_at"                   
#>