Extract table of differential results from 'K2' object

Create table differential analysis results from 'K2' object.

Usage

getDGETable(K2res)

Arguments

K2res

A K2 class object.

Value

A data.frame object with the following columns:

gene: The identifier of the gene

coef: The difference between the means of each subgroup at a given partition

mean: The mean across all observations at the given partition

t: The test statistic estimated by differential analysis

pval: The p-value estimated by differential analysis

fdr: The multiple hypothesis corrected fdr p-value, adjusted across all partitions

edge: Indication of which subgroup the gene was assigned at a given partition

node: The identifier of the partition

direction: The direction of coefficient for the assigned gene

References

Reed ER, Monti S (2021). “Multi-resolution characterization of molecular taxonomies in bulk and single-cell transcriptomics data.” Nucleic Acids Research. doi:10.1093/nar/gkab552 , https://pubmed.ncbi.nlm.nih.gov/34226941/. Ritchie ME, Phipson B, Wu D, Hu Y, Law CW, Shi W, Smyth GK (2015). “limma powers differential expression analyses for RNA-sequencing and microarray studies.” Nucleic Acids Research, 43(7), e47--e47. ISSN 1362-4962, 0305-1048, doi:10.1093/nar/gkv007 , http://academic.oup.com/nar/article/43/7/e47/2414268/limma-powers-differential-expression-analyses-for. Benjamini Y, Hochberg Y (1995). “Controlling the False Discovery Rate: A Practical and Powerful Approach to Multiple Testing.” Journal of the Royal Statistical Society: Series B (Methodological), 57(1), 289--300. ISSN 00359246, doi:10.1111/j.2517-6161.1995.tb02031.x , http://doi.wiley.com/10.1111/j.2517-6161.1995.tb02031.x.